+++ src/ads-client-subscribe.js
@@ -11,13 +11,55 @@ module.exports = function (RED) {
     this.initialDelay = config.initialDelay
     this.retryInterval = config.retryInterval
     this.controlSubscription = config.controlSubscription
+    this.resubscribeTimeout = config.resubscribeTimeout
     
     //State
+    this.subscriptionOK = false
     this.subscription = null
     this.subcribeRetryTimer = null
+    this.resubscribeTimer = null
 
     //Getting the ads-client instance
     this.connection = RED.nodes.getNode(config.connection)
+
+
+
+    //Event listeners
+    const onConnect = () => onClientStateChange('connect')
+    const onDisconnect = () => onClientStateChange('disconnect')
+    const onReconnect = () => onClientStateChange('reconnect')
+
+    /**
+     * Called when ads-client state changes
+     * @param {*} state 
+     */
+    const onClientStateChange = (state) => {
+      if (state === 'disconnect') {
+        this.subscriptionOK = false
+        this.status({ fill: 'red', shape: 'dot', text: `Error: Not connected` })
+
+      }
+      else if (state === 'reconnect') {
+
+        //Reconnected to the target
+        //If no new data received in a while -> resubscribe
+        //ads-client library should resubscribe automatically so this is just a backup if stuff goes wrong
+        //Note: wait at least resubscribeTimeout
+        this.resubscribeTimer = setTimeout(async () => {
+          if (this.subscription) {
+            const target = this.subscription ? this.subscription.target : null
+
+            this.warn(`Warning: Connection was lost and subscription didn't continue in time limit. Resubscribing to ${target}`, { target })
+
+            await unsubscribe()
+            await subscribe(target)
+          }
+        }, Math.max(this.resubscribeTimeout ? this.resubscribeTimeout : 2000, this.initialDelay * 2 + this.cycleTime * 2))
+
+      }
+    }
+
+
 
 
 
@@ -28,8 +70,15 @@ module.exports = function (RED) {
      * @param {*} sub 
      */
     const onNotificationReceived = (data, sub) => {
+      clearTimeout(this.resubscribeTimer)
+
       //Updating the subscription reference just in case (might change when PLC software changes)
       this.subscription = sub
+
+      if (!this.subscriptionOK) {
+        this.subscriptionOK = true
+        this.status({ fill: 'green', shape: 'dot', text: 'Subscribed' })
+      }
 
       //Out we go
       this.send({
@@ -50,14 +99,16 @@ module.exports = function (RED) {
 
       
       if (!this.connection) {
-        this.status({ fill: 'red', shape: 'ring', text: `Error: No connection configured` })
+        this.status({ fill: 'red', shape: 'dot', text: `Error: No connection configured` })
         this.error(`Error: No connection configured`, { error: `No connection configured` })
+        this.subscriptionOK = false
         return
       }
 
       if (this.variableName === '' && !target) {
-        this.status({ fill: 'red', shape: 'ring', text: `Error: Input msg.topic not valid string` })
+        this.status({ fill: 'red', shape: 'dot', text: `Error: Input msg.topic not valid string` })
         this.error(`Error: Input msg.topic is missing or it's not valid string`, { error: `Input msg.topic is missing or it's not valid string` })
+        this.subscriptionOK = false
         return
       }
 
@@ -68,8 +119,9 @@ module.exports = function (RED) {
 
         } catch (err) {
           //Failed to connect, we can't work..
-          this.status({ fill: 'red', shape: 'ring', text: `Error: Not connected, retrying...` })
+          this.status({ fill: 'red', shape: 'dot', text: `Error: Not connected, retrying...` })
           this.error(`Error: Not connected to the target`, { error: `Not connected to the target` })
+          this.subscriptionOK = false
           
           //Try again soon
           this.subcribeRetryTimer = setTimeout(() => subscribe(target), this.retryInterval)
@@ -77,6 +129,8 @@ module.exports = function (RED) {
           return
         }
       }
+
+      
 
       //Subscribe if not yet subscribed
       if (!this.subscription) {
@@ -91,14 +145,23 @@ module.exports = function (RED) {
             this.initialDelay
           )
 
+          this.connection.getClient().off('connect', onConnect)
+          this.connection.getClient().off('disconnect', onDisconnect)
+          this.connection.getClient().off('reconnect', onReconnect)
+          this.connection.getClient().on('connect', onConnect)
+          this.connection.getClient().on('disconnect', onDisconnect)
+          this.connection.getClient().on('reconnect', onReconnect)
+
           //Successful
-          this.status({ fill: "green", shape: "dot", text: "Subscribed" })
+          this.status({ fill: 'green', shape: 'dot', text: 'Subscribed' })
+          this.subscriptionOK = true
 
         } catch (err) {
           const errInfo = this.connection.formatError(err)
 
-          this.status({ fill: 'red', shape: 'ring', text: `Error: Subscribe failed, retrying...` })
+          this.status({ fill: 'red', shape: 'dot', text: `Error: Subscribe failed, retrying...` })
           this.error(`Error: Subscribing to ${target ? target : this.variableName} failed: ${errInfo.message} - retrying every ${this.retryInterval} ms`, errInfo)
+          this.subscriptionOK = false
 
           //Try again soon
           this.subcribeRetryTimer = setTimeout(() => subscribe(target), this.retryInterval)
@@ -125,6 +188,7 @@ module.exports = function (RED) {
         const errInfo = this.connection.formatError(err)
 
         this.error(`Error: Unsubscribing old subscription to "${this.subscription.target}" failed: ${errInfo.message}`, errInfo)
+        this.subscriptionOK = false
 
       } finally {
         this.subscription = null
@@ -158,8 +222,9 @@ module.exports = function (RED) {
 
       //Check if given topic is valid (if no variableName given)
       if (this.variableName === '' && (!msg.topic || typeof (msg.topic) !== 'string')) {
-        this.status({ fill: 'red', shape: 'ring', text: `Error: Input msg.topic not valid string` })
+        this.status({ fill: 'red', shape: 'dot', text: `Error: Input msg.topic not valid string` })
         this.error(`Error: Input msg.topic is missing or it's not valid string`, msg)
+        this.subscriptionOK = false
 
         if (done) {
           done(new Error(`Error: Input msg.topic is missing or it's not valid string`))
@@ -220,5 +285,5 @@ module.exports = function (RED) {
     }
   }
 
-  RED.nodes.registerType("ads-client-subscribe", AdsClientSubscribe)
+  RED.nodes.registerType('ads-client-subscribe', AdsClientSubscribe)
 }